More progress on test projects

+++ Common/Tests/MockVsTests/MockVsShell.cs
@@ -55,7 +55,7 @@ namespace Microsoft.VisualStudioTools.MockVsTests {
             throw new NotImplementedException();
         }
 
-        public int LoadUILibrary(ref Guid guidPackage, uint dwExFlags, out uint phinstOut) {
+        public int LoadUILibrary(ref Guid guidPackage, uint dwExFlags, out IntPtr phinstOut) {
             throw new NotImplementedException();
         }
